excel paste in fromFormat
moved the conversion to fromFormat.  but also added comment that format is junk - it is hard coded and doesn't match input.  so there is still more work to do.  but this change supports upconverting YY->YYYY and formatting in YYYY-MM-DD style which is a better overall match to the current default behavior